+++ packages/components/psammead-most-read/src/Title/index.stories.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs';
 import { withServicesKnob } from '@bbc/psammead-storybook-helpers';
-import MostReadTitle from './title';
+import MostReadTitle from '.';
 
 const arabicServiceDecorator = withServicesKnob({
   defaultService: 'arabic',
